Move grid utilities to separate module. This avoids circular imports later on, and makes it easier to work towards a Grid class
