+++ src/wam2layers/tracking/io.py
@@ -4,11 +4,11 @@ from typing import Literal
 
 import pandas as pd
 import xarray as xr
-import yaml
 
 import wam2layers
 from wam2layers.config import Config
 from wam2layers.reference.variables import DATA_ATTRIBUTES
+from wam2layers.tracking.grid import get_grid_info
 
 logger = logging.getLogger(__name__)
